requirements.txt
Django>=3.0
Pillow~=8.1.0
numpy~=1.20.1
djangorestframework~=3.12.2
seam_carving=1.0.0
